BJS2-61247 refactor: move character classes to 'unit' package
- Relocated `Unit`, `Archer`, `Mage`, and `Swordsman` classes from 'character' to 'unit' package.
- Updated imports across all affected files to reflect package change.
