+++ src/main/java/school/faang/task_61247/unit/Archer.java
@@ -1,4 +1,4 @@
-package school.faang.task_61247.character;
+package school.faang.task_61247.unit;
 
 public class Archer extends Unit {
     private static final int ARCHER_POWER = 25;
