+++ src/main/java/school/faang/task_61247/unit/Swordsman.java
@@ -1,4 +1,4 @@
-package school.faang.task_61247.character;
+package school.faang.task_61247.unit;
 
 public class Swordsman extends Unit {
     private static final int SWORDSMAN_POWER = 75;
